refactor: create user using UserController
